move derive_dimension to codegen, move to_snakcase out to lib
