+++ crates/diman_unit_system/src/codegen/dimension_def.rs
@@ -1,22 +1,6 @@
 use quote::quote;
-use syn::Ident;
 
 use crate::types::Defs;
-
-fn str_to_snakecase(s: &str) -> String {
-    let s = s.chars().rev().collect::<String>();
-    let words = s.split_inclusive(|c: char| c.is_uppercase());
-    words
-        .map(|word| word.chars().rev().collect::<String>().to_lowercase())
-        .rev()
-        .collect::<Vec<_>>()
-        .join("_")
-}
-
-pub fn to_snakecase(dim: &proc_macro2::Ident) -> Ident {
-    let snake_case = str_to_snakecase(&dim.to_string());
-    Ident::new(&snake_case, dim.span())
-}
 
 impl Defs {
     pub(crate) fn dimension_impl(&self) -> proc_macro::TokenStream {
@@ -214,18 +198,5 @@ impl Defs {
 
             }
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn str_to_snakecase() {
-        assert_eq!(super::str_to_snakecase("MyType"), "my_type".to_owned());
-        assert_eq!(super::str_to_snakecase("My"), "my".to_owned());
-        assert_eq!(
-            super::str_to_snakecase("MyVeryLongType"),
-            "my_very_long_type".to_owned()
-        );
     }
 }
